refactor(search): extracts SearchEntry into separate types file
