+++ server/src/modules/search/__tests__/search.service.spec.ts
@@ -1,7 +1,8 @@
 import { ResponseError } from '@opensearch-project/opensearch/lib/errors'
 import { StatusCodes } from 'http-status-codes'
 import { Mocker } from '../opensearch-mock'
-import { SearchEntry, SearchService } from '../search.service'
+import { SearchService } from '../search.service'
+import { SearchEntry } from '../search.types'
 
 // // Uncomment to test with live opensearch service
 // import { baseConfig, Environment } from '../../../bootstrap/config/base'
