+++ server/src/modules/search/search.types.ts
@@ -1,0 +1,8 @@
+export type SearchEntry = {
+  postId: number
+  title?: string
+  description?: string | null
+  answers?: string[]
+  agencyId?: number
+  topicId?: number | null
+}
